chore: migrate auxiliary methods to utils
